+++ src/MetaheuristicsPlatform/Algorithms/HyperHeuristics/ReinforcementLearningGreatDelugeHyperHeuristic/ReinforcementLearningGreatDelugeHyperHeuristicOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.HyperHeuristics.ReinforcementLearningGreatDelugeHyperHeuristic;
+
+public static class ReinforcementLearningGreatDelugeHyperHeuristicOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Ozcan, Misir, Ochoa & Burke",
+            2010,
+            "A Reinforcement Learning - Great-Deluge Hyper-Heuristic for Examination Timetabling",
+            "International Journal of Applied Metaheuristic Computing 1(1), 39-59",
+            "10.4018/jamc.2010102603");
+}